requirements.txt
Django~=4.2.8
djangorestframework~=3.14.0
pip~=21.2.4
emoji~=2.9.0
wheel~=0.37.0
pillow~=10.2.0
tornado~=6.2
Jinja2~=3.1.2
setuptools~=58.0.4
packaging~=22.0
zipp~=3.17.0
future~=0.18.2
platformdirs~=2.6.2
dicttoxml~=1.7.16